requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-multipart==0.0.9
yt-dlp==2024.5.27
python-dotenv==1.0.1
playwright==1.44.0